requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiomysql==0.2.0
aioredis==2.0.1
aiosignal==1.3.2
aiosmtplib==3.0.2
alembic==1.12.1
alibabacloud-credentials==1.0.2
alibabacloud-credentials-api==1.0.0
alibabacloud-dysmsapi20170525==4.2.0
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.4.1
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_util==0.3.13
alipay-sdk-python==3.7.623
annotated-types==0.7.0
anthropic==0.54.0
anyio==3.7.1
APScheduler==3.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.3.0
azure-core==1.34.0
azure-storage-blob==12.25.1
bcrypt==4.3.0
beautifulsoup4==4.14.1
blinker==1.9.0
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==44.0.3
darabonba-core==1.0.3
dashscope==1.23.4
dataclasses-json==0.6.7
DBUtils==3.1.2
distro==1.9.0
dnspython==2.7.0
dramatiq==1.17.0
ecdsa==0.19.1
email_validator==2.2.0
environs==9.5.0
exceptiongroup==1.3.0
fastapi==0.104.1
fastapi-mail==1.5.0
frozenlist==1.7.0
fuzzywuzzy==0.18.0
greenlet==3.2.3
grpcio==1.60.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-anthropic==0.3.7
langchain-community==0.3.17
langchain-core==0.3.65
langchain-deepseek==0.1.2
langchain-openai==0.3.8
langchain-text-splitters==0.3.8
langgraph==0.2.70
langgraph-checkpoint==2.0.26
langgraph-sdk==0.1.70
langsmith==0.3.45
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
minio==7.2.15
multidict==6.4.4
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.86.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pika==1.3.2
prometheus_client==0.22.1
propcache==0.3.2
protobuf==6.31.1
psycopg2==2.9.11
pyahocorasick==2.3.0
pyarrow==20.0.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.0
pydantic-settings==2.9.1
pydantic_core==2.33.0
pymilvus==2.4.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.1
redis==5.0.1
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
scipy==1.15.3
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.23
starlette==0.27.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
tushare==1.4.24
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
pyahocorasick==2.3.0
DBUtils==3.1.2
# pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple